🧬 Move the common functions or objects about operations with file to another test utility module.
