+++ test/integration_test/model/init.py
@@ -8,7 +8,8 @@ import pytest
 from pymock_api import APIConfig
 from pymock_api.model import load_config
 
-from .._spec import MockAPI_Config_Yaml_Path, run_test, yaml_factory
+from ..._file_utils import MockAPI_Config_Yaml_Path, yaml_factory
+from .._spec import run_test
 
 # [(under_test_path, expected_path)]
 DIVIDING_YAML_PATHS: List[tuple] = []
